[🛠️ REFACTOR] startGame함수 모듈화
