+++ handtris/src/components/TetrisPlay.tsx
@@ -8,7 +8,7 @@ import { playSoundEffect } from "@/hook/howl";
 import { getRoomCode } from "@/util/getRoomCode";
 import { AnimatePresence, motion } from "framer-motion";
 import { useToast } from "@/components/ui/use-toast";
-import { HandLandmarkResults, TetrisBoard } from "@/types";
+import { HandLandmarkResults } from "@/types";
 import WaitingModal from "@/components/WaitingModal";
 import LeftJoystickModel from "@/components/LeftJoystickModel";
 import RightJoystickModel from "@/components/RightJoystickModel";
@@ -18,13 +18,13 @@ import ConfettiExplosion from "react-confetti-explosion";
 import { ArrowUpNarrowWide, Donut, FlipVertical2 } from "lucide-react";
 import { LandmarkList } from "@mediapipe/hands";
 import { drawNextBlock } from "./drawNextBlock";
-import { showCountdown } from "./showCountdown";
 import { useHandleGesture } from "@/hook/useHandleGesture";
 import { updateStatus } from "@/services/gameService";
 import useFetchRoomPlayers from "@/hook/fetchRoomPlayers";
 import useSubscribeToEntering from "@/hook/useSubscribeToEntering";
 import useSubscribeToState from "@/hook/useSubscribeToState";
 import { TETRIS_CANVAS } from "@/styles";
+import { startGame } from "@/hook/startGame";
 
 const Home: React.FC = () => {
   const { toggleMusic, stopAllMusic } = useMusic();
@@ -266,122 +266,34 @@ const Home: React.FC = () => {
     }
   };
 
-  const startGame = async () => {
-    if (canvasTetrisRef.current) {
-      const ctx = canvasTetrisRef.current.getContext("2d");
-      if (ctx) {
-        ctx.clearRect(
-          0,
-          0,
-          canvasTetrisRef.current.width,
-          canvasTetrisRef.current.height,
-        );
-      }
-    }
-    if (canvasTetris2Ref.current) {
-      const ctx2 = canvasTetris2Ref.current.getContext("2d");
-      if (ctx2) {
-        ctx2.clearRect(
-          0,
-          0,
-          canvasTetris2Ref.current.width,
-          canvasTetris2Ref.current.height,
-        );
-      }
-    }
-    setIsDangerous(false);
-    setShowWaitingModal(false);
-    await new Promise(resolve => setTimeout(resolve, 800));
-    const roomCode = getRoomCode();
-    if (!handsManagerRef.current) {
-      handsManagerRef.current = new HandGestureManager(
-        setLeftHandLandmarks,
-        setRightHandLandmarks,
-        handleGesture,
-        onResults,
-      );
-      handsManagerRef.current.start(videoRef.current!);
-    }
-    await showCountdown();
-    if (canvasTetrisRef.current && canvasTetris2Ref.current) {
-      const ctx = canvasTetrisRef.current.getContext("2d")!;
-      const ctx2 = canvasTetris2Ref.current.getContext("2d")!;
-
-      try {
-        if (!isSub.current) {
-          wsManagerRef.current?.subscribe(
-            `/user/queue/tetris/${roomCode}`,
-            (message: {
-              board: TetrisBoard;
-              isEnd: boolean;
-              isAddAttack: boolean;
-              isFlipAttack: boolean;
-              isDonutAttack: boolean;
-            }) => {
-              if (tetrisGameRef.current) {
-                if (message.isEnd) {
-                  tetrisGameRef.current.gameEnd = true;
-                  stopAllMusic();
-                  playSoundEffect("/sound/winner.mp3");
-                  setGameResult("you WIN!");
-                }
-                if (message.isAddAttack) {
-                  tetrisGameRef.current.isAddAttacked = true;
-                } else if (message.isFlipAttack) {
-                  setIsFlipping(true);
-                  const playOppTetrisElement =
-                    document.getElementById("tetris-container");
-                  if (
-                    playOppTetrisElement &&
-                    !playOppTetrisElement.classList.contains("flipped-canvas")
-                  ) {
-                    playOppTetrisElement.classList.add("flipped-canvas");
-                    setTimeout(() => {
-                      setIsFlipping(false);
-                      setTimeout(() => {
-                        playOppTetrisElement.classList.add("unflipped-canvas");
-                        setTimeout(() => {
-                          playOppTetrisElement.classList.remove(
-                            "flipped-canvas",
-                          );
-                          playOppTetrisElement.classList.remove(
-                            "unflipped-canvas",
-                          );
-                        }, 500);
-                      }, 100);
-                    }, 2900);
-                  }
-                } else if (message.isDonutAttack) {
-                  tetrisGameRef.current.isDonutAttacked = true;
-                }
-                tetrisGameRef.current.drawBoard2(message.board);
-              }
-            },
-          );
-          isSub.current = true;
-        }
-
-        tetrisGameRef.current = new TetrisGame(
-          ctx,
-          ctx2,
-          wsManagerRef.current!,
-          setGameResult,
-        );
-        setLinesCleared(tetrisGameRef.current.linesCleared);
-        tetrisGameRef.current.roomCode = getRoomCode();
-      } catch (error) {
-        console.error("Failed to connect to WebSocket for game", error);
-      }
-    }
-    toggleMusic();
-  };
   useSubscribeToState(
     wsManagerRef,
     isOwner,
     setIsAllReady,
     setIsReady,
     setIsStart,
-    startGame,
+    () =>
+      startGame({
+        canvasTetrisRef,
+        canvasTetris2Ref,
+        setIsDangerous,
+        setShowWaitingModal,
+        setLeftHandLandmarks,
+        setRightHandLandmarks,
+        handleGesture,
+        onResults,
+        videoRef,
+        wsManagerRef,
+        handsManagerRef,
+        tetrisGameRef,
+        setLinesCleared,
+        setGameResult,
+        isSub,
+        setIsFlipping,
+        stopAllMusic,
+        playSoundEffect,
+        toggleMusic,
+      }),
     isStart,
   );
   useEffect(() => {
